Fix blockquote empty overhang
When the blockquote element only contains a p element but no footer
element or similar trailing tag, the ::after pseudo class paints a
trailing empty line which only contains the ">" character. This is
caused by an overhanging paragraph margin. Removing the margin for any
last element contained in blockquote removes this visual glitch.